Move callback tests.
